requirements.txt
Django==1.6
PyYAML==3.11
amqp==1.4.6
anyjson==0.3.3
argh==0.26.1
backports.ssl-match-hostname==3.4.0.2
billiard==3.3.0.19
celery==3.1.17
celerymon==1.0.3
certifi==14.05.14
django-celery==3.1.16
django-registration==1.0
django-supervisor==0.3.2
kombu==3.0.24
librabbitmq==1.6.1
meld3==1.0.0
pathtools==0.1.2
pika==0.9.14
pytz==2014.10
supervisor==3.1.3
tornado==4.0.2
watchdog==0.8.2
wsgiref==0.1.2
